open_door/ransac_package/requirements.txt
numpy 
scipy 
matplotlib 
opencv-python 
mpl_toolkits
open3d 
simplejson 
pyyaml